requirements.txt
streamlit
pynacl
cryptography